# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "INVENTRAK Inventory Management API",
    "description": "Backend API for the INVENTRAK inventory management system. Both the SQLite (Express) backend and the npm-free fallback expose this identical contract. Live docs: GET /api/docs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local backend"
    }
  ],
...
  "paths": {
...
    "/api/order-inquiries": {
      "get": {
        "tags": [
          "Order Inquiries"
        ],
        "operationId": "listOrderInquiries",
        "summary": "List order inquiries",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "fulfilled",
                "delivered"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inquiries (plain array, or { data, pagination } when page/limit provided)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OrderInquiry"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/PaginatedInquiries"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Order Inquiries"
        ],
        "operationId": "createOrderInquiry",
        "summary": "Submit an order inquiry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInquiryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inquiry submitted (with the GCash payment step when payment_method is gcash/card)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/order-inquiries/{id}": {
      "put": {
        "tags": [
          "Order Inquiries"
        ],
        "operationId": "updateOrderInquiry",
        "summary": "Update inquiry status (pending -> approved -> fulfilled / rejected)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "approved",
                      "rejected",
                      "fulfilled",
                      "delivered"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inquiry updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/order-inquiries/{id}/costing": {
      "get": {
        "tags": [
          "Order Inquiries"
        ],
        "operationId": "getInquiryCosting",
        "summary": "Get the immutable costing snapshot frozen when the inquiry was submitted",
        "description": "Written once at submission and never rewritten, so the profit a customer was quoted does not drift when the catalog is repriced. Scoped like the inquiry: admins see any, a customer sees only their own.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Costing snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostingRecord"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/order-inquiries/{id}/payment": {
      "put": {
        "tags": [
          "Order Inquiries"
        ],
        "operationId": "updateInquiryPayment",
        "summary": "Mark an inquiry as paid/unpaid/failed after the GCash step",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed or malformed input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Admin access required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CostingRecord": {
        "type": "object",
        "required": [
          "inquiry_id",
          "total_revenue",
          "target_quantity",
          "cost_basis",
          "lines_priced",
          "lines_total",
          "margin_percent",
          "computed_at"
        ],
        "properties": {
          "inquiry_id": {
            "type": "integer",
            "description": "The order inquiry this snapshot belongs to"
          },
          "total_cost": {
            "type": "number",
            "nullable": true,
            "description": "Sum of unit cost x quantity. Null when no line could be costed."
          },
          "total_revenue": {
            "type": "number",
            "description": "What the customer was charged at submission (the inquiry's estimated_cost)."
          },
          "target_quantity": {
            "type": "number",
            "description": "Sum of line quantities \u2014 one ordered unit is one 'cup'."
          },
          "cost_per_cup": {
            "type": "number",
            "nullable": true,
            "description": "total_cost / target_quantity."
          },
          "suggested_selling_price": {
            "type": "number",
            "nullable": true,
            "description": "Revenue needed to hit margin_percent on this cost: total_cost / (1 - margin)."
          },
          "estimated_profit": {
            "type": "number",
            "nullable": true,
            "description": "total_revenue - total_cost."
          },
          "cost_basis": {
            "type": "string",
            "enum": [
              "exact",
              "imputed",
              "none"
            ],
            "description": "How much of total_cost is real: every priced line had a known cost (exact); some did and the rest were estimated from the blended cost ratio (imputed); nothing could be costed (none)."
          },
          "lines_priced": {
            "type": "integer",
            "description": "Order lines whose product had a known unit cost."
          },
          "lines_total": {
            "type": "integer",
            "description": "Total order lines."
          },
          "margin_percent": {
            "type": "number",
            "description": "Target margin used for suggested_selling_price (COSTING_TARGET_MARGIN, default 30)."
          },
          "computed_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "notify": {
            "type": "object",
            "description": "Delivery status of the verification code (present only when a code was actually generated: register + resend-verification). email is always attempted; sms only when a phone was provided.",
            "properties": {
              "email": {
                "type": "boolean"
              },
              "sms": {
                "type": "boolean"
              }
            }
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
...
      "PaginatedInquiries": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderInquiry"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "OrderInquiryInput": {
        "type": "object",
        "required": [
          "customer_name",
          "customer_email"
        ],
        "properties": {
          "customer_name": {
            "type": "string",
            "maxLength": 100
          },
          "customer_email": {
            "type": "string",
            "maxLength": 100
          },
          "customer_phone": {
            "type": "string",
            "maxLength": 20,
            "description": "Mobile number for SMS status updates (optional)"
          },
          "delivery_address": {
            "type": "string",
            "maxLength": 500,
            "description": "Delivery address collected at checkout (optional)"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "cod",
              "gcash",
              "card",
              "other"
            ],
            "default": "cod",
            "description": "Payment method chosen at checkout"
          },
          "products": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "object",
                  "required": [
                    "name",
                    "qty"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "integer",
                      "nullable": true
                    },
                    "name": {
                      "type": "string",
                      "maxLength": 200,
                      "description": "Product name as the customer ordered it"
                    },
                    "qty": {
                      "type": "integer",
                      "minimum": 1,
                      "description": "Quantity ordered"
                    },
                    "price": {
                      "type": "number",
                      "minimum": 0,
                      "description": "Request field: unit price actually charged (deal price when a flash pick)"
                    },
                    "original_price": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true,
                      "description": "Pre-discount catalog price for deal items (null when no discount)"
                    },
                    "unit_price": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true,
                      "description": "Stored/normalized unit price the customer pays (null for legacy string lines)"
                    },
                    "subtotal": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true,
                      "description": "Stored line total (unit_price x qty, null when unpriced)"
                    }
                  }
                }
              ]
            },
            "description": "Ordered lines: legacy strings (\"Widget x2\") or structured objects with prices"
          },
          "estimated_cost": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
      "OrderInquiry": {
        "type": "object",
        "required": [
          "id",
          "customer_name",
          "customer_email",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "customer_name": {
            "type": "string"
          },
          "customer_email": {
            "type": "string"
          },
          "customer_phone": {
            "type": "string",
            "nullable": true
          },
          "customer_id": {
            "type": "integer",
            "nullable": true,
            "description": "Links to the Customer Record this order belongs to. Guests get a Customer Record too \u2014 they simply have no account. Null only when the order carries no usable identity at all."
          },
          "products": {
            "type": "string"
          },
          "estimated_cost": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "delivery_address": {
            "type": "string",
            "nullable": true
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "cod",
              "gcash",
              "card",
              "other"
            ],
            "default": "cod"
          },
          "payment_status": {
            "type": "string",
            "enum": [
              "unpaid",
              "paid",
              "failed"
            ],
            "default": "unpaid",
            "description": "GCash/card payment state (COD inquiries stay unpaid)"
          },
          "payment_reference": {
            "type": "string",
            "nullable": true,
            "description": "GCash-style payment reference code"
          },
          "payment_url": {
            "type": "string",
            "nullable": true,
            "description": "PayMongo checkout URL (real gateway) when configured"
          },
          "payment_qr": {
            "type": "string",
            "nullable": true,
            "description": "QR image URL the customer scans to pay"
          },
          "payment_provider": {
            "type": "string",
            "nullable": true,
            "description": "paymongo | demo"
          },
          "user_id": {
            "type": "integer",
            "nullable": true,
            "description": "Account that placed the inquiry (null for legacy/guest orders)"
          },
          "status_history": {
            "type": "string",
            "nullable": true,
            "description": "JSON array of { status, at } progress events (Placed -> Approved -> Delivered)"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "fulfilled",
              "delivered"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "products_detail": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "object",
                  "required": [
                    "name",
                    "qty"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "integer",
                      "nullable": true
                    },
                    "name": {
                      "type": "string",
                      "maxLength": 200,
                      "description": "Product name as the customer ordered it"
                    },
                    "qty": {
                      "type": "integer",
                      "minimum": 1,
                      "description": "Quantity ordered"
                    },
                    "price": {
                      "type": "number",
                      "minimum": 0,
                      "description": "Request field: unit price actually charged (deal price when a flash pick)"
                    },
                    "original_price": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true,
                      "description": "Pre-discount catalog price for deal items (null when no discount)"
                    },
                    "unit_price": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true,
                      "description": "Stored/normalized unit price the customer pays (null for legacy string lines)"
                    },
                    "subtotal": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true,
                      "description": "Stored line total (unit_price x qty, null when unpriced)"
                    }
                  }
                }
              ]
            },
            "description": "Parsed line items with per-line prices (unit_price, original_price, subtotal)"
          }
        }
      },
      "InquiryCreated": {
        "type": "object",
        "required": [
          "ok",
          "message",
          "id"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "description": "Id of the created inquiry"
          },
          "payment": {
            "type": "object",
            "description": "Payment step for GCash/card checkout (present when payment_method is gcash or card)",
            "required": [
              "payment_method",
              "payment_status",
              "payment_reference",
              "payment_qr"
            ],
            "properties": {
              "payment_method": {
                "type": "string"
              },
              "payment_status": {
                "type": "string"
              },
              "payment_reference": {
                "type": "string"
              },
              "payment_url": {
                "type": "string",
                "nullable": true
              },
              "payment_qr": {
                "type": "string"
              }
            }
          }
        }
      },
      "PaymentUpdate": {
        "type": "object",
        "required": [
          "payment_status"
        ],
        "properties": {
          "payment_status": {
            "type": "string",
            "enum": [
              "paid",
              "unpaid",
              "failed"
            ]
          }
        }
      },
...
    }
  }
}